+++ layouts/company/single.html
@@ -108,7 +108,15 @@
 
     {{ partialCached "related-companies.html" . .Params.categories (.Param "relevant-countries") }}
 
-    <div id="comments-widget" data-rating-enabled="1" data-display-warning="1"></div>
+    <!--TODO: no styling?-->
+    <noscript class="noscript noscript-comments">{{ T "noscript-comments" | safeHTML }}</noscript>
+    {{ if .Resources.GetMatch "comments.ssr.html"}}
+    <div id="comments-widget" data-rating-enabled="1" data-display-warning="1">
+        <h2>Comments</h2><!-- Translate this? But this is in i18n/comments, but I need it in i18n/hugo-->
+        {{ (.Resources.GetMatch "comments.ssr.html").Content | safeHTML}}</div>
+    {{ end }}
+
+
 </main>
 {{ end }}
