+++ package.json
@@ -44,6 +44,7 @@
     "homepage": "https://www.datenanfragen.de",
     "dependencies": {
         "autocomplete.js": "^0.37.0",
+        "axios": "^0.22.0",
         "brutusin-json-forms": "https://github.com/brutusin/json-forms",
         "js-cookie": "^2.2.1",
         "letter-generator": "^1.0.0",
@@ -57,6 +58,8 @@
         "qrcode": "^1.4.4",
         "react-joyride": "^1.11.4",
         "react-scrollspy": "^3.4.0",
+        "ts-node": "^10.2.1",
+        "typescript": "^4.4.3",
         "typesense": "^0.3.0"
     },
     "devDependencies": {
@@ -100,6 +103,7 @@
         "postcss-loader": "^3.0.0",
         "postcss-preset-env": "^6.7.0",
         "postcss-url": "^8.0.0",
+        "preact-render-to-string": "^5.1.19",
         "prettier": "^2.0.5",
         "speed-measure-webpack-plugin": "^1.3.1",
         "stylelint": "^13.4.1",
