+++ src/i18n/de.json
@@ -111,6 +111,7 @@
         "noscript-my-requests": "Alle Anfragen werden mithilfe von JavaScript auf deinem Computer gespeichert (außer natürlich JavaScript ist deaktiviert). Der Code wird keine Daten an andere Server weiterleiten, wenn Du das nicht möchstest. Wenn Du Anfragen speichern und ansehen möchtest musst du JavaScript im Browser aktivieren oder eine Ausnahme in Deine Blocker hinzufügen. Sollte es dann immer noch nicht funktionieren, schreib uns eine E-Mail an <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a>.",
         "noscript-wizard": "An dieser Stelle befände sich ein mit JavaScript generierter Wizard, zur Erstellung von Anfragen auf Datenauskunft. Die Erstellung der Anfragen benötigt JavaScript. Du kannst die Seite benutzen, ohne JavaScript zu aktiveren, hast dann aber leider nur Zugriff auf eingeschränkte Funktionen. Der Code wird keine Daten an andere Server weiterleiten, wenn Du das nicht möchtest. Wenn Du die Webseite im vollem Umfang nutzen möchtest musst du JavaScript im Browser aktivieren oder eine Ausnahme in Deine Blocker hinzufügen. Sollte es dann immer noch nicht funktionieren, schreib uns eine E-Mail an <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a>.",
         "noscript-actwidget": "An dieser Stelle befände sich ein Helfer, mit dem Du schnell eine DSGVO-Anfrage generieren kannst. Er läuft nur auf deinem Computer und verwendet deshalb ausschließlich JavaScript. Wenn Du ihn benutzen willst, dann aktiviere bitte JavaScript in deinem Browser. Sollte es dann immer noch nicht funktionieren, schreib uns eine E-Mail an <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a>.",
+        "noscript-comments": "Schalte JavaScript ein, um Kommentare schreiben zu können.",
         "404-title": "Fehler 404: Seite nicht gefunden",
         "404-body": "Leider existiert die gesuchte Seite auf diesem Server nicht. Vielleicht hast du dich ja einfach in der URL vertippt? Falls der Fehler wieder auftritt, hinterlasse uns doch einen Hinweis auf unserer E-Mail-Adresse <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a> oder schreib in unseren <a href='https://github.com/datenanfragen/website/issues/new'>Issue-Tracker</a>.",
         "404-alt": "Eine Person die auf der 0 eines großen 404-Schildes sitzt. Sie hält ein Buch, in dem eine Seite fehlt.",
