+++ src/i18n/en.json
@@ -112,6 +112,7 @@
         "noscript-my-requests": "All requests are saved and shown on your computer using JavaScript (of course only if it is activated). The code won't send any data to any servers if you do not consent. If you want to save and see your requests on your computer, please activate JavaScript in your browser or add an exception to your blocker. If it still doesn't work after that, send us an email at <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a>.",
         "noscript-wizard": "Here would have been a wizard that is generated by JavaScript, which helps you creating your requests. Creating requests only on your computer needs JavaScript. You can use the website without JavaScript but you only have limited functionality. The code won't send any data to any servers if you do not consent. If you want to use the website and all its features, please activate JavaScript in your browser or add an exception to your blocker. If it still doesn't work after that, send us an email at <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a>.",
         "noscript-actwidget": "Here would have been a tool to quickly generate a GDPR request. It only runs on your computer and is therefore completely written in JavaScript. If you want to use it, please enable JavaScript in your browser. If it still doesn't work after that, send us an email at <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a>.",
+        "noscript-comments": "Enable JavaScript to write comments",
         "press-no-releases": "At this time there are no press releases in English.",
         "404-title": "Error 404: Page not found",
         "404-body": "The page you were looking for sadly doesn't exist on this server. Maybe you just have a typo in the URL? If the error persists, perhaps notify us via our email <a href='mailto:dev@datenanfragen.de'>dev@datenanfragen.de</a> or submit an issue in our <a href='https://github.com/datenanfragen/website/issues/new'>issue tracker</a>.",
